+++ db/sqlc/models.go
@@ -11,23 +11,23 @@ import (
 )
 
 type Account struct {
-	ID        int64     `json:"id"`
-	OwnerID   int64     `json:"owner_id"`
-	Balance   int64     `json:"balance"`
+	ID        int32     `json:"id"`
+	OwnerID   int32     `json:"owner_id"`
+	Balance   int32     `json:"balance"`
 	Currency  string    `json:"currency"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
 type Entry struct {
-	ID        int64     `json:"id"`
-	AccountID int64     `json:"account_id"`
-	Amount    int64     `json:"amount"`
+	ID        int32     `json:"id"`
+	AccountID int32     `json:"account_id"`
+	Amount    int32     `json:"amount"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
 type Session struct {
 	ID           uuid.UUID `json:"id"`
-	UserID       int64     `json:"user_id"`
+	UserID       int32     `json:"user_id"`
 	RefreshToken string    `json:"refresh_token"`
 	UserAgent    string    `json:"user_agent"`
 	ClientIp     string    `json:"client_ip"`
@@ -37,15 +37,15 @@ type Session struct {
 }
 
 type Transfer struct {
-	ID            int64     `json:"id"`
-	FromAccountID int64     `json:"from_account_id"`
-	ToAccountID   int64     `json:"to_account_id"`
-	Amount        int64     `json:"amount"`
+	ID            int32     `json:"id"`
+	FromAccountID int32     `json:"from_account_id"`
+	ToAccountID   int32     `json:"to_account_id"`
+	Amount        int32     `json:"amount"`
 	CreatedAt     time.Time `json:"created_at"`
 }
 
 type User struct {
-	ID                int64     `json:"id"`
+	ID                int32     `json:"id"`
 	Username          string    `json:"username"`
 	HashedPassword    string    `json:"hashed_password"`
 	FullName          string    `json:"full_name"`
